Extract ClickCanvas to a non-public function in entry tests
