+++ test/test.go
@@ -79,13 +79,6 @@ func AssertRendersToMarkup(t *testing.T, masterFilename string, c fyne.Canvas, m
 	return true
 }
 
-// ClickCanvas simulates a primary mouse button click at an absolute position on the canvas.
-func ClickCanvas(c fyne.Canvas, pos fyne.Position) {
-	if o, p := findMouseable(c, pos); o != nil {
-		clickPrimary(c, o.(desktop.Mouseable), &fyne.PointEvent{AbsolutePosition: pos, Position: p})
-	}
-}
-
 // Drag drags at an absolute position on the canvas.
 // deltaX/Y is the dragging distance: <0 for dragging up/left, >0 for dragging down/right.
 func Drag(c fyne.Canvas, pos fyne.Position, deltaX, deltaY float32) {
@@ -279,15 +272,6 @@ func findTappable(c fyne.Canvas, pos fyne.Position) (o fyne.CanvasObject, p fyne
 	return
 }
 
-func findMouseable(c fyne.Canvas, pos fyne.Position) (o fyne.CanvasObject, p fyne.Position) {
-	matches := func(object fyne.CanvasObject) bool {
-		_, ok := object.(desktop.Mouseable)
-		return ok
-	}
-	o, p, _ = driver.FindObjectAtPositionMatching(pos, matches, c.Overlays().Top(), c.Content())
-	return
-}
-
 func prepareTap(obj interface{}, pos fyne.Position) (*fyne.PointEvent, fyne.Canvas) {
 	d := fyne.CurrentApp().Driver()
 	ev := &fyne.PointEvent{Position: pos}
@@ -302,19 +286,6 @@ func prepareTap(obj interface{}, pos fyne.Position) (*fyne.PointEvent, fyne.Canv
 func tap(c fyne.Canvas, obj fyne.Tappable, ev *fyne.PointEvent) {
 	handleFocusOnTap(c, obj)
 	obj.Tapped(ev)
-}
-
-func clickPrimary(c fyne.Canvas, obj desktop.Mouseable, ev *fyne.PointEvent) {
-	handleFocusOnTap(c, obj)
-	mouseEvent := &desktop.MouseEvent{
-		PointEvent: *ev,
-		Button:     desktop.MouseButtonPrimary,
-	}
-	obj.MouseDown(mouseEvent)
-	obj.MouseUp(mouseEvent)
-	if tap, ok := obj.(fyne.Tappable); ok {
-		tap.Tapped(ev)
-	}
 }
 
 func handleFocusOnTap(c fyne.Canvas, obj interface{}) {
